Moving shared LLVM utils. Cleaning up params
