+++ qirlib/src/lib.rs
@@ -20,6 +20,8 @@ extern crate llvm_sys_140 as llvm_sys;
 #[cfg(not(feature = "no-llvm-linking"))]
 pub mod builder;
 #[cfg(not(feature = "no-llvm-linking"))]
+pub(crate) mod llvm_utils;
+#[cfg(not(feature = "no-llvm-linking"))]
 pub mod qis;
 #[cfg(not(feature = "no-llvm-linking"))]
 pub mod rt;
